Split some sbt build sources
For later re-use as-is by Mill
