+++ project/Build.scala
@@ -1,6 +1,5 @@
 import java.io.File
 import java.nio.file._
-import ScaladocGeneration._
 import com.jsuereth.sbtpgp.PgpKeys
 import sbt.Keys.*
 import sbt.*
@@ -13,10 +12,12 @@ import com.typesafe.sbt.packager.windows.WindowsPlugin
 import com.typesafe.sbt.packager.windows.WindowsPlugin.autoImport.Windows
 import sbt.Package.ManifestAttributes
 import sbt.PublishBinPlugin.autoImport._
+import dotty.tools.sbtplugin._
 import dotty.tools.sbtplugin.RepublishPlugin
 import dotty.tools.sbtplugin.RepublishPlugin.autoImport._
 import dotty.tools.sbtplugin.ScalaLibraryPlugin
 import dotty.tools.sbtplugin.ScalaLibraryPlugin.autoImport._
+import dotty.tools.sbtplugin.ScaladocGeneration._
 import dotty.tools.sbtplugin.DottyJSPlugin
 
 import sbt.ScriptedPlugin.autoImport._
@@ -35,115 +36,11 @@ import org.portablescala.sbtplatformdeps.PlatformDepsPlugin.autoImport._
 
 import sbt.dsl.LinterLevel.Ignore
 
+import Constants._
+
 object Build {
   import ScaladocConfigs._
-
-  /** Version of the Scala compiler used to build the artifacts.
-   *  Reference version should track the latest version pushed to Maven:
-   *  - In main branch it should be the last RC version
-   *  - In release branch it should be the last stable release
-   *
-   *  Warning: Change of this variable needs to be consulted with `expectedTastyVersion`
-   */
-  val referenceVersion = "3.8.4-RC3"
-
-  /** Version of the Scala compiler targeted in the current release cycle
-   *  Contains a version without RC/SNAPSHOT/NIGHTLY specific suffixes
-   *  Should be updated ONLY after release or cutoff for previous release cycle.
-   *
-   *  Should only be referred from `dottyVersion` or settings/tasks requiring simplified version string,
-   *  eg. `compatMode` or Windows native distribution version.
-   *
-   *  Warning: Change of this variable might require updating `expectedTastyVersion`
-   */
-  val developedVersion = "3.10.0"
-
-  /** The version of the compiler including the RC prefix.
-   *  Defined as common base before calculating environment specific suffixes in `dottyVersion`
-   *
-   *  By default, during development cycle defined as `${developedVersion}-RC1`;
-   *  During release candidate cycle incremented by the release officer before publishing a subsequent RC version;
-   *  During final, stable release is set exactly to `developedVersion`.
-  */
-  val baseVersion = s"$developedVersion-RC1"
-
-  /** The version of TASTY that should be emitted, checked in runtime test
-   *  For details on how TASTY version should be set see related discussions:
-   *    - https://github.com/scala/scala3/issues/13447#issuecomment-912447107
-   *    - https://github.com/scala/scala3/issues/14306#issuecomment-1069333516
-   *    - https://github.com/scala/scala3/pull/19321
-   *
-   *  Simplified rules, given 3.$minor.$patch = $developedVersion
-   *    - Major version is always 28
-   *    - TASTY minor version:
-   *      - in main (NIGHTLY): {if $patch == 0 || ${referenceVersion.matches(raw"3.$minor.0-RC\d")} then $minor else ${minor + 1}}
-   *      - in release branch is always equal to $minor
-   *    - TASTY experimental version:
-   *      - in main (NIGHTLY) is always experimental
-   *      - in release candidate branch is experimental if {patch == 0}
-   *      - in stable release is always non-experimental
-   */
-  val expectedTastyVersion = "28.10-experimental-1"
-  checkReleasedTastyVersion()
-
-  /** Final version of Scala compiler, controlled by environment variables. */
-  val dottyVersion = {
-    if (isRelease) baseVersion
-    else if (isNightly) {
-      val formatter = java.time.format.DateTimeFormatter.ofPattern("yyyyMMdd")
-      val currentDate =
-        formatter.format(java.time.ZonedDateTime.now(java.time.ZoneId.of("UTC")))
-      s"${baseVersion}-bin-${currentDate}-${VersionUtil.gitHash}-NIGHTLY"
-    } else if (isBenchmark) {
-      s"${baseVersion}-bin-${VersionUtil.gitHashFull}-BENCH"
-    }
-    else s"${baseVersion}-bin-SNAPSHOT"
-  }
-  def isRelease = sys.env.get("RELEASEBUILD").contains("yes")
-  def isNightly = sys.env.get("NIGHTLYBUILD").contains("yes")
-  def isBenchmark = sys.env.get("BENCHMARKBUILD").contains("yes")
-
-  /** Version calculated for `nonbootstrapped` projects */
-  val dottyNonBootstrappedVersion = {
-    // Make sure sbt always computes the scalaBinaryVersion correctly
-    val bin = if (!dottyVersion.contains("-bin")) "-bin" else ""
-    dottyVersion + bin + "-nonbootstrapped"
-  }
-
-  // LTS or Next
-  val versionLine = "Next"
-
-  /** Minor version against which we check binary compatibility.
-   *
-   *  This must be the earliest published release in the same versioning line.
-   *  For a developedVersion `3.M.P` the mimaPreviousDottyVersion should be set to:
-   *   - `3.M.0`     if `P > 0`
-   *   - `3.(M-1).0` if `P = 0`
-   */
-  val mimaPreviousDottyVersion = "3.8.0" // TODO: update to 3.9.0 when released
-
-  /** Version of Scala CLI to download */
-  val scalaCliLauncherVersion = "1.14.0"
-  /** Version of Coursier to download for initializing the local maven repo of Scala command */
-  val coursierJarVersion = "2.1.25-M25"
-
-  object CompatMode {
-    final val BinaryCompatible = 0
-    final val SourceAndBinaryCompatible = 1
-  }
-
-  val compatMode = {
-    val VersionRE = """^\d+\.(\d+)\.(\d+)""".r
-    developedVersion match {
-      case VersionRE(_, "0")   => CompatMode.BinaryCompatible
-      case _                   => CompatMode.SourceAndBinaryCompatible
-    }
-  }
-
-  val homepageUrl = "https://scala-lang.org/"
-  val dottyOrganization = "org.scala-lang"
-  val dottyGithubUrl = "https://github.com/scala/scala3"
-  val dottyGithubRawUserContentUrl = "https://raw.githubusercontent.com/scala/scala3"
+  import Versions._
 
   // Run tests with filter through vulpix test suite
   val testCompilation = inputKey[Unit]("runs integration test with the supplied filter")
@@ -309,13 +206,6 @@ object Build {
 
   private lazy val currentYear: String = java.util.Calendar.getInstance().get(java.util.Calendar.YEAR).toString
 
-  private val shellBanner: String =
-    """%n      ________ ___   / /  ___
-       |%n    / __/ __// _ | / /  / _ |
-       |%n  __\\ \\/ /__/ __ |/ /__/ __ |
-       |%n /____/\\___/_/ |_/____/_/ | |
-       |%n                          |/  %s""".stripMargin.replace("\n", "")
-
   // Common generator for properties files
   lazy val generatePropertiesFile = (fileName: String, contents: Def.Initialize[String]) => Def.task {
     val file = (Compile / resourceManaged).value / fileName
@@ -348,42 +238,9 @@ object Build {
   lazy val generateLibraryProperties: Def.Initialize[Task[Seq[File]]] = {
     val fileName = "library.properties"
     val contents = Def.setting {
-      s"""version.number=${version.value}
-          |maven.version.number=${version.value}
-          |copyright.string=Copyright 2002-$currentYear, LAMP/EPFL
-          |shell.banner=${shellBanner}
-          |""".stripMargin
+      LibraryProperties.content(version.value, currentYear)
     }
     generatePropertiesFile(fileName, contents)
-  }
-
-  def scalacOptionsDocSettings(includeExternalMappings: Boolean = true) = {
-    val extMap = Seq("-external-mappings:" +
-        (if (includeExternalMappings) ".*scala/.*::scaladoc3::https://nightly.scala-lang.org/api/," else "") +
-        ".*java/.*::javadoc::https://docs.oracle.com/javase/8/docs/api/")
-    Seq(
-      "-skip-by-regex:.+\\.internal($|\\..+)",
-      "-skip-by-regex:.+\\.impl($|\\..+)",
-      "-project-logo", "docs/_assets/images/logo.svg",
-      "-social-links:" +
-        "github::https://github.com/scala/scala3," +
-        "discord::https://discord.com/invite/scala," +
-        "twitter::https://twitter.com/scala_lang",
-      // contains special definitions which are "transplanted" elsewhere
-      // and which therefore confuse Scaladoc when accessed from this pkg
-      "-skip-by-id:scala.runtime.stdLibPatches",
-      // MatchCase is a special type that represents match type cases,
-      // Reflect doesn't expect to see it as a standalone definition
-      // and therefore it's easier just not to document it
-      "-skip-by-id:scala.runtime.MatchCase",
-      "-skip-by-id:dotty.tools.tasty",
-      "-skip-by-id:dotty.tools.tasty.util",
-      "-skip-by-id:dotty.tools.tasty.besteffort",
-      "-project-footer", s"Copyright (c) 2002-$currentYear, LAMP/EPFL",
-      "-author",
-      "-groups",
-      "-default-template", "static-site-main"
-    ) ++ extMap
   }
 
   val enableBspAllProjects = sys.env.get("ENABLE_BSP_ALL_PROJECTS").map(_.toBoolean).getOrElse{
@@ -470,7 +327,7 @@ object Build {
       assert(docScalaInstance.loaderCompilerOnly == base.loaderCompilerOnly)
       docScalaInstance
     },
-    Compile / doc / scalacOptions ++= scalacOptionsDocSettings(),
+    Compile / doc / scalacOptions ++= ScaladocOptions.scalacOptionsDocSettings(),
   )
 
   // Settings used when compiling dotty with a non-bootstrapped dotty
@@ -541,7 +398,7 @@ object Build {
       (thisProjectID.organization % crossedName % mimaPreviousDottyVersion)
     },
 
-    mimaCheckDirection := (compatMode match {
+    mimaCheckDirection := (CompatMode.value match {
       case CompatMode.BinaryCompatible          => "backward"
       case CompatMode.SourceAndBinaryCompatible => "both"
     }),
@@ -568,69 +425,6 @@ object Build {
   /** Like `findArtifact` but returns the absolute path of the entry as a string */
   def findArtifactPath(classpath: Def.Classpath, name: String): String =
     findArtifact(classpath, name).getAbsolutePath
-
-  /** Replace package names in package definitions, for shading.
-   * It assumes the full package def is written on a single line.
-   * It does not adapt the imports accordingly.
-   */
-  def replacePackage(lines: List[String])(replace: PartialFunction[String, String]): List[String] = {
-    def recur(lines: List[String]): List[String] =
-      lines match {
-        case head :: tail =>
-          if (head.startsWith("package ")) {
-            val packageName = head.stripPrefix("package ").trim
-            val newPackageName = replace.applyOrElse(packageName, (_: String) => packageName)
-            s"package $newPackageName" :: tail
-          } else head :: recur(tail)
-        case _ => lines
-      }
-    recur(lines)
-  }
-
-  /** Insert UnsafeNulls Import after package */
-  def insertUnsafeNullsImport(lines: List[String]): List[String] = {
-    def recur(ls: List[String], foundPackage: Boolean): List[String] = ls match {
-      case l :: rest =>
-        val lt = l.trim()
-        if (foundPackage) {
-          if (!(lt.isEmpty || lt.startsWith("package ")))
-            "import scala.language.unsafeNulls" :: ls
-          else l :: recur(rest, foundPackage)
-        } else {
-          if (lt.startsWith("package ")) l +: recur(rest, true)
-          else l :: recur(rest, foundPackage)
-        }
-      case _ => ls
-    }
-    recur(lines, false)
-  }
-
-  /** replace imports of `com.google.protobuf.*` with compiler implemented version */
-  def replaceProtobuf(lines: List[String]): List[String] = {
-    def recur(ls: List[String]): List[String] = ls match {
-      case l :: rest =>
-        val lt = l.trim()
-        if (lt.isEmpty || lt.startsWith("package ") || lt.startsWith("import ")) {
-          val newLine =
-            if (lt.startsWith("import com.google.protobuf.")) {
-              if (lt == "import com.google.protobuf.CodedInputStream") {
-                "import dotty.tools.dotc.semanticdb.internal.SemanticdbInputStream as CodedInputStream"
-              } else if (lt == "import com.google.protobuf.CodedOutputStream") {
-                "import dotty.tools.dotc.semanticdb.internal.SemanticdbOutputStream as CodedOutputStream"
-              } else {
-                l
-              }
-            } else {
-              l
-            }
-          newLine :: recur(rest)
-        } else {
-          ls // don't check rest of file
-        }
-      case _ => ls
-    }
-    recur(lines)
-  }
 
   def insertClasspathInArgs(args: List[String], cp: String): List[String] = {
     val (beforeCp, fromCp) = args.span(_ != "-classpath")
@@ -956,7 +750,7 @@ object Build {
       scriptedBatchExecution := true,
       scriptedLaunchOpts ++= Seq(
         s"-Dplugin.scalaVersion=${dottyVersion}",
-        s"-Dplugin.scala2Version=${ScalaLibraryPlugin.scala2Version}",
+        s"-Dplugin.scala2Version=${Versions.scala2Version}",
         s"-Dplugin.scalaJSVersion=${scalaJSVersion}",
       ),
       scriptedBufferLog := true,
@@ -1164,7 +958,7 @@ object Build {
     .settings(
       name          := "scala2-library",
       moduleName    := "scala2-library",
-      scalaVersion  := ScalaLibraryPlugin.scala2Version,
+      scalaVersion  := Versions.scala2Version,
       version       := scalaVersion.value,
       // Remove Scala 3 specific settings
       scalacOptions --= Seq(
@@ -1678,10 +1472,10 @@ object Build {
           val sjsSources = (trgDir ** "*.scala").get.toSet
           sjsSources.foreach(f => {
             val lines = IO.readLines(f)
-            val linesWithPackage = replacePackage(lines) {
+            val linesWithPackage = Shading.replacePackage(lines) {
               case "org.scalajs.ir" => "dotty.tools.sjs.ir"
             }
-            IO.writeLines(f, insertUnsafeNullsImport(linesWithPackage))
+            IO.writeLines(f, Shading.insertUnsafeNullsImport(linesWithPackage))
           })
           sjsSources
         } (Set(scalaJSIRSourcesJar)).toSeq
@@ -1802,10 +1596,10 @@ object Build {
           val sjsSources = (trgDir ** "*.scala").get.toSet
           sjsSources.foreach(f => {
             val lines = IO.readLines(f)
-            val linesWithPackage = replacePackage(lines) {
+            val linesWithPackage = Shading.replacePackage(lines) {
               case "org.scalajs.ir" => "dotty.tools.sjs.ir"
             }
-            IO.writeLines(f, insertUnsafeNullsImport(linesWithPackage))
+            IO.writeLines(f, Shading.insertUnsafeNullsImport(linesWithPackage))
           })
           sjsSources
         } (Set(scalaJSIRSourcesJar)).toSeq
@@ -1946,65 +1740,10 @@ object Build {
         // It would be the easiest to create a temp directory and apply patches there, but this task would be used frequently during development
         // If we'd build using copy the emitted warnings would point developers to copies instead of original sources. Any fixes made in there would be lost.
         // Instead let's apply revertable patches to the files as part snapshot doc generation process
-        abstract class SourcePatch(val file: File) {
-          def apply(): Unit
-          def revert(): Unit
-        }
         val docs = file("docs")
         val sourcePatches = if (justAPI) Nil else Seq(
-          // Generate full sidebar.yml based on template and reference content
-          new SourcePatch(docs / "sidebar.yml") {
-            val referenceSideBarCopy = IO.temporaryDirectory / "sidebar.yml.copy"
-            IO.copyFile(file, referenceSideBarCopy)
-
-            override def apply(): Unit = {
-              val yaml = new org.yaml.snakeyaml.Yaml()
-              type YamlObject = java.util.Map[String, AnyRef]
-              type YamlList[T] = java.util.List[T]
-              def loadYaml(file: File): YamlObject = {
-                val reader = Files.newBufferedReader(file.toPath)
-                try yaml.load(reader).asInstanceOf[YamlObject]
-                finally reader.close()
-              }
-              // Ensure to always operate on original (Map, List) instances
-              val template = loadYaml(docs / "sidebar.nightly.template.yml")
-              template.get("subsection")
-                .asInstanceOf[YamlList[YamlObject]]
-                .stream()
-                .filter(_.get("title") == "Reference")
-                .findFirst()
-                .orElseThrow(() => new IllegalStateException("Reference subsection not found in sidebar.nightly.template.yml"))
-                .putAll(loadYaml(referenceSideBarCopy))
-
-              val sidebarWriter = Files.newBufferedWriter(this.file.toPath)
-              try yaml.dump(template, sidebarWriter)
-              finally sidebarWriter.close()
-            }
-            override def revert(): Unit = IO.move(referenceSideBarCopy, file)
-          },
-          // Add patch about nightly version usage
-          new SourcePatch(docs / "_layouts" / "static-site-main.html") {
-            lazy val originalContent = IO.read(file)
-
-            val warningMessage = """{% if page.nightlyOf %}
-              |  <aside class="warning">
-              |    <div class='icon'></div>
-              |    <div class='content'>
-              |      This is a nightly documentation. The content of this page may not be consistent with the current stable version of language.
-              |      Click <a href="{{ page.nightlyOf }}">here</a> to find the stable version of this page.
-              |    </div>
-              |  </aside>
-              |{% endif %}""".stripMargin
-
-            override def apply(): Unit = {
-              IO.write(file,
-                originalContent
-                .replace("{{ content }}", s"$warningMessage {{ content }}")
-                .ensuring(_.contains(warningMessage), "patch to static-site-main layout not applied!")
-              )
-            }
-            override def revert(): Unit = IO.write(file, originalContent)
-          }
+          ScaladocGeneration.sidebarSourcePatch(docs),
+          ScaladocGeneration.nightlyVersionNoteSourcePatch(docs)
         )
         val config = Def.task {
           outputDirOverride
@@ -2060,7 +1799,7 @@ object Build {
           .add(Revision(version))
           .add(OutputDir(s"scaladoc/output/${version}"))
           .add(SiteRoot(docs.getAbsolutePath))
-          .add(defaultSourceLinks(version = version, allowGitSHA = false))
+          .add(ScaladocConfigs0.defaultSourceLinks(version = version, allowGitSHA = false))
           .remove[ApiSubdirectory]
         }
         generateDocumentation(config)
@@ -2078,41 +1817,20 @@ object Build {
 
         // Move all the source files to a temporary directory and apply some changes specific to the reference documentation
         val docs = IO.createTemporaryDirectory
-        IO.copyDirectory(file("docs"), docs)
-        IO.delete(docs / "_blog")
 
-        // Add redirections from previously supported URLs, for some pages
-        for (name <- Seq("changed-features", "contextual", "dropped-features", "metaprogramming", "other-new-features")) {
-          val path = docs / "_docs" / "reference" / name / s"${name}.md"
-          val contentLines = IO.read(path).linesIterator.to[collection.mutable.ArrayBuffer]
-          contentLines.insert(1, s"redirectFrom: /${name}.html") // Add redirection
-          val newContent = contentLines.mkString("\n")
-          IO.write(path, newContent)
-        }
+        ScaladocGeneration.prepareReferenceDocumentationDir(docs)
 
-        val outputDir = "scaladoc/output/reference"
         val languageReferenceConfig = Def.task {
-          Scala3.value
-            .add(OutputDir(outputDir))
-            .add(SiteRoot(docs.getAbsolutePath))
-            .add(ProjectName("Scala 3 Reference"))
-            .add(ProjectVersion(baseVersion))
-            .remove[VersionsDictionaryUrl]
-            .add(SourceLinks(List(
-              s"${docs.getAbsolutePath}=github://scala/scala3/language-reference-stable#docs"
-            )))
-            .add(GenerateAPI(false))
-            .add(SnippetCompiler(referenceSnippetCompilerTargets(docs.getAbsolutePath)))
+          ScaladocGeneration.languageReferenceConfig(docs, Scala3.value)
         }
 
         val generateDocs = generateDocumentation(languageReferenceConfig)
 
         val expectedLinksRegeneration = Def.task {
           if (shouldRegenerateExpectedLinks) {
-            val script = (file("project") / "scripts" / "regenerateExpectedLinks").toString
-            val expectedLinksFile = (file("project") / "scripts" / "expected-links" / "reference-expected-links.txt").toString
+            val command = ScaladocGeneration.expectedLinksRegenerationCommand()
             import _root_.scala.sys.process._
-            s"$script $outputDir $expectedLinksFile".!
+            command.mkString(" ").!
           }
         }
 
@@ -2125,7 +1843,7 @@ object Build {
         val referenceRoot = docsRoot / "_docs" / "reference"
         val requested =
           if (selected.nonEmpty) selected
-          else referenceSnippetRelativeRoots
+          else ScaladocConfigs0.referenceSnippetRelativeRoots
 
         val tempRoot = IO.createTemporaryDirectory
         val tempDocsRoot = tempRoot / "_docs" / "reference"
@@ -2183,7 +1901,7 @@ object Build {
               .add(NoLinkWarnings(true))
               .add(NoLinkAssetWarnings(true))
               .add(GenerateAPI(false))
-              .add(SnippetCompiler(referenceSnippetCompilerTargets(tempRoot.getAbsolutePath)))
+              .add(SnippetCompiler(ScaladocConfigs0.referenceSnippetCompilerTargets(tempRoot.getAbsolutePath)))
           }
 
           generateDocumentation(config)
@@ -2223,7 +1941,6 @@ object Build {
       BuildInfoPlugin.buildInfoDefaultSettings
 
   lazy val presentationCompilerSettings = {
-    val mtagsVersion = "1.6.7"
     Seq(
       libraryDependencies ++= Seq(
         "org.lz4" % "lz4-java" % "1.8.1",
@@ -2231,7 +1948,7 @@ object Build {
         "org.scalameta" % "mtags-interfaces" % mtagsVersion,
         "com.google.guava" % "guava" % "33.6.0-jre",
       ),
-      libraryDependencies += ("org.scalameta" % s"mtags-shared_${ScalaLibraryPlugin.scala2Version}" % mtagsVersion % SourceDeps),
+      libraryDependencies += ("org.scalameta" % s"mtags-shared_${Versions.scala2Version}" % mtagsVersion % SourceDeps),
       ivyConfigurations += SourceDeps.hide,
       transitiveClassifiers := Seq("sources"),
       publishLocal := publishLocal.dependsOn( // It is best to publish all together. It is not rare to make changes in both compiler / presentation compiler and it can get misaligned
@@ -2266,7 +1983,7 @@ object Build {
           val mtagsSharedSources = (targetDir ** "*.scala").get.toSet
           mtagsSharedSources.foreach(f => {
             val lines = IO.readLines(f)
-            val substitutions = (replaceProtobuf(_)) andThen (insertUnsafeNullsImport(_))
+            val substitutions = (Shading.replaceProtobuf(_)) andThen (Shading.insertUnsafeNullsImport(_))
             IO.writeLines(f, substitutions(lines))
           })
           mtagsSharedSources
@@ -2936,46 +2653,6 @@ object Build {
 
   }
 
-  /* Tests TASTy version invariants during NIGHLY, RC or Stable releases */
-  def checkReleasedTastyVersion(): Unit = {
-    case class ScalaVersion(minor: Int, patch: Int, isRC: Boolean)
-    def parseScalaVersion(version: String): ScalaVersion = version.split("\\.|-").take(4) match {
-      case Array("3", minor, patch)    => ScalaVersion(minor.toInt, patch.toInt, false)
-      case Array("3", minor, patch, _) => ScalaVersion(minor.toInt, patch.toInt, true)
-      case other => sys.error(s"Invalid Scala base version string: $baseVersion")
-    }
-    lazy val version = parseScalaVersion(baseVersion)
-    lazy val referenceV = parseScalaVersion(referenceVersion)
-    lazy val (tastyMinor, tastyIsExperimental) = expectedTastyVersion.split("\\.|-").take(4) match {
-      case Array("28", minor)                    => (minor.toInt, false)
-      case Array("28", minor, "experimental", _) => (minor.toInt, true)
-      case other => sys.error(s"Invalid TASTy version string: $expectedTastyVersion")
-    }
-
-    if(isNightly) {
-      assert(tastyIsExperimental, "TASTY needs to be experimental in nightly builds")
-      val expectedTastyMinor = version.patch match {
-        case 0 => version.minor
-        case 1 if referenceV.patch == 0 && referenceV.isRC =>
-          // Special case for a period when reference version is a new unstable minor
-          // Needed for non_bootstrapped tests requiring either stable tasty or the same experimental version produced by both reference and bootstrapped compiler
-          assert(version.minor == referenceV.minor, "Expected reference and base version to use the same minor")
-          version.minor
-        case _ => version.minor + 1
-      }
-      assert(tastyMinor == expectedTastyMinor, "Invalid TASTy minor version")
-    }
-
-    if(isRelease) {
-      assert(version.minor == tastyMinor, "Minor versions of TASTY vesion and Scala version should match in release builds")
-      assert(!referenceV.isRC, "Stable release needs to use stable compiler version")
-      if (version.isRC && version.patch == 0)
-        assert(tastyIsExperimental, "TASTy should be experimental when releasing a new minor version RC")
-      else
-        assert(!tastyIsExperimental, "Stable version cannot use experimental TASTY")
-    }
-  }
-
   /** Helper to validate JAR contents */
   private def validateJarIsEmpty(jar: File): File = {
     val jarFile = new java.util.jar.JarFile(jar)
@@ -3017,131 +2694,30 @@ object Build {
 
 object ScaladocConfigs {
   import Build._
-  private lazy val currentYear: String = java.util.Calendar.getInstance().get(java.util.Calendar.YEAR).toString
-
-  def dottyExternalMapping = ".*scala/.*::scaladoc3::https://nightly.scala-lang.org/api/"
-  def javaExternalMapping = ".*java/.*::javadoc::https://docs.oracle.com/en/java/javase/17/docs/api/java.base/"
-  def defaultSourceLinks(version: String, allowGitSHA: Boolean = true) = {
-    def dottySrcLink(v: String) = sys.env.get("GITHUB_SHA") match {
-      case Some(sha) if allowGitSHA => s"github://scala/scala3/$sha"
-      case None => s"github://scala/scala3/$v"
-    }
-    SourceLinks(List(dottySrcLink(version), "docs=github://scala/scala3/main#docs"))
-  }
 
   lazy val DefaultGenerationSettings = Def.task {
-    def projectVersion = version.value
-    def socialLinks = SocialLinks(List(
-      "github::https://github.com/scala/scala3",
-      "discord::https://discord.com/invite/scala",
-      "twitter::https://twitter.com/scala_lang",
-    ))
-    def projectLogo = ProjectLogo("docs/_assets/images/logo.svg")
-    def skipByRegex = SkipByRegex(List(".+\\.internal($|\\..+)", ".+\\.impl($|\\..+)"))
-    def skipById = SkipById(List(
-      "scala.runtime.stdLibPatches",
-      "scala.runtime.MatchCase",
-    ))
-    def projectFooter = ProjectFooter(s"Copyright (c) 2002-$currentYear, LAMP/EPFL")
-    def defaultTemplate = DefaultTemplate("static-site-main")
-    GenerationConfig(
-      List(),
-      ProjectVersion(projectVersion),
-      GenerateInkuire(true),
-      defaultSourceLinks(version = dottyVersion),
-      skipByRegex,
-      skipById,
-      projectLogo,
-      socialLinks,
-      projectFooter,
-      defaultTemplate,
-      Author(true),
-      Groups(true),
-      QuickLinks(
-        List(
-          "Learn::https://docs.scala-lang.org/",
-          "Install::https://www.scala-lang.org/download/",
-          "Playground::https://scastie.scala-lang.org",
-          "Find\u00A0A\u00A0Library::https://index.scala-lang.org",
-          "Community::https://www.scala-lang.org/community/",
-          "Blog::https://www.scala-lang.org/blog/",
-        )
-      )
-    )
+    ScaladocConfigs0.DefaultGenerationSettings(version.value)
   }
 
   lazy val Scaladoc = Def.task {
-    DefaultGenerationSettings.value
-      .add(UseJavacp(true))
-      .add(ProjectName("scaladoc"))
-      .add(OutputDir("scaladoc/output/self"))
-      .add(Revision(VersionUtil.gitHash))
-      .add(ExternalMappings(List(dottyExternalMapping, javaExternalMapping)))
-      .withTargets((Compile / classDirectory).value.getAbsolutePath :: Nil)
+    ScaladocConfigs0.Scaladoc(
+      version.value,
+      (Compile / classDirectory).value
+    )
   }
 
   lazy val Testcases = Def.task {
     val tastyRoots = (Test / Build.testcasesOutputDir).value
-    DefaultGenerationSettings.value
-      .add(UseJavacp(true))
-      .add(OutputDir("scaladoc/output/testcases"))
-      .add(ProjectName("scaladoc testcases"))
-      .add(Revision("main"))
-      .add(SnippetCompiler(List("scaladoc-testcases/docs=compile")))
-      .add(SiteRoot("scaladoc-testcases/docs"))
-      .add(CommentSyntax(List(
-        "scaladoc-testcases/src/example/comment-md=markdown",
-        "scaladoc-testcases/src/example/comment-wiki=wiki"
-      )))
-      .add(ExternalMappings(List(dottyExternalMapping, javaExternalMapping)))
-      .withTargets(tastyRoots)
+    ScaladocConfigs0.Testcases(version.value, tastyRoots)
   }
 
-  def snippetCompilerTargets(dottyLibSrc:String) = List(
-    s"$dottyLibSrc/scala=compile"
-  )
-  // Relative subtrees in `_docs/reference` where snippet compilation is explicitly enabled.
-  // Keep this shared with the full docs tasks and the lightweight snippet-check task.
-  def referenceSnippetRelativeRoots = List(
-    "new-types",
-    "enums",
-    "experimental/capture-checking",
-  )
-  def captureCheckingSnippetTestTargets(docsRoot: String) = List(
-    s"$docsRoot/_docs/reference/experimental/capture-checking/basics.md=compile+test",
-    s"$docsRoot/_docs/reference/experimental/capture-checking/checked-exceptions.md=compile+test",
-    s"$docsRoot/_docs/reference/experimental/capture-checking/scoped-capabilities.md=compile+test"
-  )
-  def referenceSnippetCompilerTargets(docsRoot: String) =
-    referenceSnippetRelativeRoots.map(path => s"$docsRoot/_docs/reference/$path=compile") ++
-      captureCheckingSnippetTestTargets(docsRoot)
-
   lazy val Scala3 = Def.task {
-    val stdlib = { // relative path to the stdlib directory ('library/')
-      val projectRoot = (ThisBuild/baseDirectory).value.toPath
-      val stdlibRoot = (`scala-library-bootstrapped` / baseDirectory).value
-      projectRoot.relativize(stdlibRoot.toPath.normalize())
-    }
-
-    DefaultGenerationSettings.value
-      .add(ProjectName("Scala 3"))
-      .add(OutputDir(file("scaladoc/output/scala3").getAbsoluteFile.getAbsolutePath))
-      .add(Revision("main"))
-      .add(ExternalMappings(List(javaExternalMapping)))
-      .add(DocRootContent((stdlib / "resources" / "rootdoc.txt").toString))
-      .add(CommentSyntax(List(
-        // Markdown syntax is used by default for all sources
-        "markdown"
-      )))
-      .add(VersionsDictionaryUrl("https://scala-lang.org/api/versions.json"))
-      .add(DocumentSyntheticTypes(true))
-      .add(SnippetCompiler(
-        snippetCompilerTargets(s"$stdlib/src") ++ referenceSnippetCompilerTargets("docs")
-      ))
-      .add(NoSnippetNamesFor(List("_docs/reference")))
-      .add(SiteRoot("docs"))
-      .add(ApiSubdirectory(true))
-      .withTargets((`scala-library-bootstrapped` / Compile / products).value.map(_.getAbsolutePath))
+    ScaladocConfigs0.Scala3(
+      version.value,
+      (ThisBuild/baseDirectory).value.toPath,
+      (`scala-library-bootstrapped` / baseDirectory).value.toPath,
+      (`scala-library-bootstrapped` / Compile / products).value.map(_.getAbsolutePath)
+    )
   }
 
 }
