+++ project/Constants.scala
@@ -1,0 +1,7 @@
+
+object Constants {
+  val homepageUrl = "https://scala-lang.org/"
+  val dottyOrganization = "org.scala-lang"
+  val dottyGithubUrl = "https://github.com/scala/scala3"
+  val dottyGithubRawUserContentUrl = "https://raw.githubusercontent.com/scala/scala3"
+}
