+++ project/DocumentationWebsite.scala
@@ -2,7 +2,6 @@ import java.io.File
 import java.net.URI
 import java.nio.file.Paths
 import sbt._
-import Build._
 
 object DocumentationWebsite {
   def generateStaticAssets(
