+++ project/MiMaFilters.scala
@@ -1,3 +1,4 @@
+package dotty.tools.sbtplugin
 
 import com.typesafe.tools.mima.core._
 
@@ -7,7 +8,7 @@ object MiMaFilters {
 
     val ForwardsBreakingChanges: Map[String, Seq[ProblemFilter]] = Map(
       // Additions that require a new minor version of the library
-      Build.mimaPreviousDottyVersion -> Seq(
+      Versions.mimaPreviousDottyVersion -> Seq(
         ProblemFilters.exclude[DirectMissingMethodProblem]("scala.caps.package#package.freeze"),
         ProblemFilters.exclude[DirectMissingMethodProblem]("scala.caps.package#package.cap"),
         ProblemFilters.exclude[FinalClassProblem]("scala.Function1$UnliftOps$"),
@@ -23,7 +24,7 @@ object MiMaFilters {
       // Only exceptional cases should be added here.
 
       // Breaking changes since last reference version
-      Build.mimaPreviousDottyVersion -> Seq(
+      Versions.mimaPreviousDottyVersion -> Seq(
         ProblemFilters.exclude[DirectMissingMethodProblem]("scala.None.orNull"),
         ProblemFilters.exclude[MissingTypesProblem]("scala.util.control.NonLocalReturns$ReturnThrowable"),
         // THIS IS FINE, IT SHOULD HAVE BEEN THIS WAY
@@ -60,7 +61,7 @@ object MiMaFilters {
   object ScalaLibrarySJS {
     val ForwardsBreakingChanges: Map[String, Seq[ProblemFilter]] = Map(
       // Additions that require a new minor version of the library
-      Build.mimaPreviousDottyVersion -> Seq(
+      Versions.mimaPreviousDottyVersion -> Seq(
         // No .class files generated in the artifacts, only `scala.scalajs.*` files might be present
         ProblemFilters.exclude[MissingClassProblem]("scala.*"),
       ),
@@ -68,26 +69,26 @@ object MiMaFilters {
 
     val BackwardsBreakingChanges: Map[String, Seq[ProblemFilter]] = Map(
       // We should never break backwards compatibility
-      Build.mimaPreviousDottyVersion -> Seq.empty,
+      Versions.mimaPreviousDottyVersion -> Seq.empty,
     )
   }
 
   object TastyCore {
     val ForwardsBreakingChanges: Map[String, Seq[ProblemFilter]] = Map(
       // Additions that require a new minor version of tasty core
-      Build.mimaPreviousDottyVersion -> Seq(
+      Versions.mimaPreviousDottyVersion -> Seq(
       )
     )
 
     val BackwardsBreakingChanges: Map[String, Seq[ProblemFilter]] = Map(
-      Build.mimaPreviousDottyVersion -> Seq.empty,
+      Versions.mimaPreviousDottyVersion -> Seq.empty,
     )
   }
 
   object Interfaces {
     val ForwardsBreakingChanges: Map[String, Seq[ProblemFilter]] = Map(
       // Additions that require a new minor version of interfaces
-      Build.mimaPreviousDottyVersion -> Seq.empty,
+      Versions.mimaPreviousDottyVersion -> Seq.empty,
     )
 
     val BackwardsBreakingChanges: Map[String, Seq[ProblemFilter]] = Map.empty
