+++ project/RepublishPlugin.scala
@@ -343,62 +343,13 @@ object RepublishPlugin extends AutoPlugin {
   }
 
   private def generateVersionFile() = Def.task[Unit] {
-    import scala.util.Try
-    import java.time.format.DateTimeFormatterBuilder
-    import java.time.format.SignStyle
-    import java.time.temporal.ChronoField.*
-    import java.time.ZoneId
-    import java.time.Instant
-    import java.time.ZonedDateTime
-    import java.time.ZonedDateTime
-    import java.util.Locale
-    import java.util.Date
-
     val base: File = new File(".") // Using the working directory as base for readability
     val s = streams.value
     val log = s.log
     val progVersion = version.value
     val distDir = republishRepo.value
 
-    def write(path: String, content: String) {
-      val p = distDir / path
-      IO.write(p, content)
-    }
-
-    val humanReadableTimestampFormatter = new DateTimeFormatterBuilder()
-        .parseCaseInsensitive()
-        .appendValue(YEAR, 4, 10, SignStyle.EXCEEDS_PAD)
-        .appendLiteral('-')
-        .appendValue(MONTH_OF_YEAR, 2)
-        .appendLiteral('-')
-        .appendValue(DAY_OF_MONTH, 2)
-        .appendLiteral(' ')
-        .appendValue(HOUR_OF_DAY, 2)
-        .appendLiteral(':')
-        .appendValue(MINUTE_OF_HOUR, 2)
-        .appendLiteral(':')
-        .appendValue(SECOND_OF_MINUTE, 2)
-        .appendOffset("+HHMM", "Z")
-        .toFormatter(Locale.US)
-
-    // Retrieve build time
-    val systemZone = ZoneId.systemDefault().normalized()
-    val timestamp  = ZonedDateTime.ofInstant(Instant.ofEpochMilli(new Date().getTime), systemZone)
-    val buildTime  = humanReadableTimestampFormatter.format(timestamp)
-
-    // Check the current Git revision
-    val gitRevision: String = Try {
-      if ((base / ".git").exists()) {
-        log.info("[republish] Checking the git revision of the current project")
-        sys.process.Process("git rev-parse HEAD").!!
-      } else {
-        "unknown"
-      }
-    }.getOrElse("unknown").trim
-
-
-    // Output the version number and Git revision
-    write("VERSION", s"version:=${progVersion}\nrevision:=${gitRevision}\nbuildTime:=${buildTime}\n")
+    Dist.generateVersionFile(base, distDir, progVersion, log.info(_))
   }
 
   override val projectSettings: Seq[Def.Setting[_]] = Def.settings(
