+++ project/VersionUtil.scala
@@ -1,3 +1,5 @@
+package dotty.tools.sbtplugin
+
 import org.eclipse.jgit.storage.file.FileRepositoryBuilder
 import org.eclipse.jgit.lib.{Constants, ObjectId, Ref, Repository}
 import org.eclipse.jgit.revwalk.{RevCommit, RevWalk}
